remove pandas dependancy & move method to utils.py
