+++ requirements-dev.txt
@@ -1,3 +1,2 @@
 prospector>=1.3.0
 pre-commit
-deepdiff
