+++ requirements.txt
@@ -4,7 +4,6 @@ dash_daq
 plotly>=4
 cryptography>=2.6
 Werkzeug>=1.0.0
-pandas
 oauth2_client
 requests
 pytz
@@ -16,6 +15,7 @@ geojson
 reverse_geocode
 androguard
 pycryptodomex
+deepdiff
 
 #swagger req
 certifi >= 14.05.14
